statistics moved to new command group
