+++ src/vss_tools/cli.py
@@ -49,15 +49,28 @@ def cli(ctx: click.Context, log_level: str, log_file: Path):
         "tree": "vss_tools.exporters.tree:cli",
         "samm": "vss_tools.exporters.samm:cli",
         "go": "vss_tools.exporters.go:cli",
-        "statisticsradial": "vss_tools.exporters.statisticsradial:cli",
-        "statisticssankey": "vss_tools.exporters.statisticssankey:cli",
-        "statisticspiechart": "vss_tools.exporters.statisticspiechart:cli",
     },
 )
 @click.pass_context
 def export(ctx: click.Context):
     """
     Export a vspec to a chosen format
+    """
+    if ctx.invoked_subcommand is None:
+        click.echo(ctx.get_help())
+
+@cli.group(
+    cls=LazyGroup,
+    lazy_subcommands={
+        "radial": "vss_tools.stats.statisticsradial:cli",
+        "sankey": "vss_tools.stats.statisticssankey:cli",
+        "piechart": "vss_tools.stats.statisticspiechart:cli",
+    },
+)
+@click.pass_context
+def stats(ctx: click.Context):
+    """
+    Export statistics to a chosen format
     """
     if ctx.invoked_subcommand is None:
         click.echo(ctx.get_help())
